feat(ui5-li-notification, ui5-li-notification-group): migrate to TypeScript
