+++ packages/fiori/src/NotificationListGroupItem.hbs
@@ -5,7 +5,7 @@
 	@keydown="{{_onkeydown}}"
 	role="listitem"
 	tabindex="{{_tabIndex}}"
-	aria-expanded="{{ariaExpanded}}"
+	aria-expanded="{{isAriaExpanded}}"
 	aria-labelledby="{{ariaLabelledBy}}"
 >
 	<div class="ui5-nli-group-header">
